+++ Arkanoid/src/app/BonusEnlarge.java
@@ -10,7 +10,7 @@ import javax.imageio.ImageIO;
 
 public class BonusEnlarge extends Bonus {
     //ESTE BONUS ALARGA LA NAVE
-    public BonusEnlarge(Escenario escenario,int x,int y){
+    public BonusEnlarge(Escenario escenario,double x,double y){
         this.x=x;
         this.y=y;
         this.cuerpo=new Rectangle();
@@ -32,8 +32,8 @@ public class BonusEnlarge extends Bonus {
     }
 
     @Override
-    public int velocidad() {
-        return 1;
+    public double velocidad() {
+        return 1.0;
     }
 
     @Override
@@ -84,7 +84,7 @@ public class BonusEnlarge extends Bonus {
 
     @Override
     public void draw(Graphics2D g) {
-        g.drawImage(this.imagen, this.x, this.y, null);
+        g.drawImage(this.imagen,(int) this.x,(int) this.y, null);
     }
 
     @Override
